Cleanup deprecated methods.

+++ multiplatform-crypto-libsodium-bindings/src/jsMain/kotlin/com/ionspin/kotlin/crypto/kdf/Kdf.kt
@@ -27,7 +27,7 @@ actual object Kdf {
     ): UByteArray {
         return getSodium().crypto_kdf_derive_from_key(
             subkeyLength.toUInt(),
-            subkeyId.toUInt(),
+            subkeyId,
             context,
             masterKey.toUInt8Array()
         ).toUByteArray()
